Filter for batch

+++ src/main/java/com/group2/KoiFarmShop/dto/reponse/BatchPageReponse.java
@@ -14,7 +14,8 @@ public class BatchPageReponse {
     List<BatchReponse> batchReponses;
     private int pageNum;
     private int pageSize;
-    private int totalElements;
+    private long totalElements;
+
     private int totalPages;
 
 }
